Update buzzer and irlineTracking files
- Removed irLineTracking_interrupt.c files because not using it.
- Rename irLineTracking_Timer.c file to irLineTracking.c
- Added comments and removed some commented codes in buzzer.c and irLineTracking.c
- irLineTracking.c  file -> Rename function name from void irLineTracking_UsingTimer(void) to void irLineTracking(void)
